+++ src/test/java/com/bootcamp/be_java_hisp_w25_g14/repository/PostRepoTest.java
@@ -1,0 +1,8 @@
+package com.bootcamp.be_java_hisp_w25_g14.repository;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+public class PostRepoTest {
+}
